Fix File::deflate when compressionLevel=0.
Also add Unknown115 to end of file, as newer binlog does it.
FileStatistics uses first reserved field to store original fileSize.

+++ src/Vector/BLF/File.h
@@ -200,7 +200,7 @@ public:
     int compressionLevel;
 
     /** default log container size */
-    ULONGLONG defaultLogContainerSize;
+    DWORD defaultLogContainerSize;
 
     /**
      * open file
